# api/openapi.yaml
openapi: 3.1.0
info:
  title: Squyrrl API
  version: 0.1.0
  description: |
    Squyrrl（箩筐）主 API 契约。本文件描述 Phase 1 已实现的 20 个端点。

    ## 认证模型

    - 客户端用邮箱 OTP 登录得到 `access_token`（1h）+ `refresh_token`（30d）。
    - 受保护端点要求 `Authorization: Bearer <access_token>` 请求头。
    - access token 过期后用 `/auth/refresh` 旋转出新的 access；refresh token 不旋转。
    - 任何 token 都仅以 SHA-256 哈希入库；明文 token 永不持久化。

    ## 同步并发控制

    - `snippets.version` 是单调自增的乐观锁。
    - 写入端点（PATCH / DELETE）必须携带 `If-Match: "<version>"` 头，否则 400 / 409。
    - 读取端点会返回 `ETag: "<version>"` 响应头。
    - 增量同步：`GET /snippets?since=<RFC3339>` 返回 `updated_at >= since` 的碎片，按时间升序，附带 `next_since` 用于下一次请求。
    - `?include_deleted=true` 让同步流包含软删除的碎片（带 `deleted_at`），客户端据此本地下架。

    ## 数据混淆

    - `snippets.text_content` 与未来的文件 blob 在客户端用 XChaCha20-Poly1305 + 应用静态密钥混淆后再上传。
    - 服务器只见混淆字节，不解密。在 JSON 中字段以 base64 出现（Go `[]byte` 默认编码）。
  contact:
    name: Squyrrl
    url: https://squyrrl.app
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: 开发环境（本机 Docker Compose）
  - url: https://api.squyrrl.app
    description: 生产环境（占位）
...
security:
  - bearerAuth: []

# =============================================================================
# 端点
# =============================================================================
paths:
...
  /pages:
    post:
      operationId: createPage
      tags: [Pages]
      summary: 创建页面
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PageCreateInput' }
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    get:
      operationId: listPages
      tags: [Pages]
      summary: 列出当前用户的页面
      description: 按 `order_idx` 升序、`created_at` 升序返回；不含软删除条目。
      responses:
        '200':
          description: 页面列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Page' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /pages/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    get:
      operationId: getPage
      tags: [Pages]
      summary: 拉取单个页面
      responses:
        '200':
          description: 页面
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updatePage
      tags: [Pages]
      summary: 更新页面
      description: "系统页面（`is_system: true`）禁止改名 / 改隐藏，但可调 `order_idx`。"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PageUpdateInput' }
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 系统页面只读
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deletePage
      tags: [Pages]
      summary: 软删除页面
      description: |
        软删除该页面，并把该用户下指向它的所有未删除碎片的 `page_id` 置为 NULL，
        同时把这些碎片的 `version` 加 1（保证客户端通过同步流能看到变更）。
        系统页面禁止删除。
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 系统页面只读
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Tags
...
  /files/check:
    post:
      operationId: checkFile
      tags: [Files]
      summary: 按 plain_hash 查重
      description: |
        客户端先把文件明文 SHA-256 提交，命中则直接拿到 file_id 复用，
        未命中再走 `POST /files` 上传字节。节流量并实现跨用户去重（ADR-004）。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FileCheckInput' }
      responses:
        '200':
          description: 查重结果
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FileCheckResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /files:
    post:
      operationId: uploadFile
      tags: [Files]
      summary: 上传混淆字节
      description: |
        请求体为客户端混淆后的二进制（XChaCha20-Poly1305 + 应用静态密钥）。
        必须以 header 形式申报 plain_hash / cipher_hash / 大小 / MIME；
        服务端会重新计算上传字节的 SHA-256 与 X-Cipher-Hash 比对。
      parameters:
        - in: header
          name: X-Plain-Hash
          required: true
          schema: { type: string, pattern: '^[0-9a-f]{64}$' }
        - in: header
          name: X-Cipher-Hash
          required: true
          schema: { type: string, pattern: '^[0-9a-f]{64}$' }
        - in: header
          name: X-Size-Bytes
          required: true
          schema: { type: integer, minimum: 1 }
        - in: header
          name: X-Mime
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: 已上传或已命中去重
          content:
            application/json:
              schema:
                type: object
                required: [file]
                properties:
                  file: { $ref: '#/components/schemas/File' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '413':
          description: 单文件超过 100MiB
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /files/{id}:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    get:
      operationId: getFileMetadata
      tags: [Files]
      summary: 获取文件元数据
      description: |
        仅返回 file 元数据 JSON（id / plain_hash / cipher_hash / size / mime），不传字节。
        客户端在显示已附附件预览前先调此端点拿 mime，再根据 mime 决定是否调 `download` 拉缩略图字节。
      responses:
        '200':
          description: 文件元数据
          content:
            application/json:
              schema: { $ref: '#/components/schemas/File' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /files/{id}/download:
    parameters:
      - $ref: '#/components/parameters/UUIDPath'
    get:
      operationId: downloadFile
      tags: [Files]
      summary: 下载混淆字节
      description: |
        返回客户端混淆字节流；客户端用应用静态密钥解混淆得到原文。
        响应头携带 X-Plain-Hash / X-Cipher-Hash 便于校验。
      responses:
        '200':
          description: 混淆字节
          headers:
            X-Plain-Hash:
              schema: { type: string }
            X-Cipher-Hash:
              schema: { type: string }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- URIs
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    internalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: 仅供 Bot 等内部服务调用；与主 API 共享同一个秘密
  parameters:
    UUIDPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: 请求参数无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 未认证 / token 无效
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 资源不存在或不归属当前用户
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 人可读的错误描述（中文） }

    # ---------- 认证 ----------
...
    Page:
      type: object
      required: [id, name, is_hidden, is_system, order_idx, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        is_hidden: { type: boolean }
        is_system: { type: boolean }
        order_idx: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PageCreateInput:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          description: 客户端可选指定 UUID（离线生成）
        name: { type: string, minLength: 1, maxLength: 100 }
        is_hidden: { type: boolean, default: false }
        order_idx: { type: integer, default: 0 }
    PageUpdateInput:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        is_hidden: { type: boolean }
        order_idx: { type: integer }

    # ---------- 标签 ----------
...
    File:
      type: object
      required: [id, plain_hash, cipher_hash, size_bytes, mime, created_at]
      properties:
        id: { type: string, format: uuid }
        plain_hash: { type: string, description: 明文 SHA-256 hex }
        cipher_hash: { type: string, description: 混淆字节 SHA-256 hex }
        size_bytes: { type: integer, format: int64 }
        mime: { type: string }
        created_at: { type: string, format: date-time }
    FileCheckInput:
      type: object
      required: [plain_hash]
      properties:
        plain_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
    FileCheckResponse:
      type: object
      required: [exists]
      properties:
        exists: { type: boolean }
        file:
          oneOf:
            - $ref: '#/components/schemas/File'
            - type: 'null'

    # ---------- URI 解析 ----------
...